Combine TicketList and TicketTable components
